Readded impl Send for FdbFuture
The rational for removing this was that under a pure rust rewrite it might not be possible for the Future to be Send. This is fair, but right now lots of the code in this library is tightly coupled to the C driver anyway. In a hypothetical rewrite people would have to change their code a ton anyway.